Adding a 'webapps' group for common permissions
This adds a conventional group 'webapps' using the new 'system-groups' role.
This group is used as the primary role for team members to ensure files created by a member can be edited by others with suitable group permissions.

Note: The 'app' user is currently not a member of this group, however I am open to changing this in future.